move namespaceTest from AdminApiTest because it init multiple broker and accessing bundleCache can be from different broker which fails test
